refacto main
